requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
streamlit==1.36.0
plotly==5.22.0
matplotlib==3.9.0
tqdm==4.66.4
pydantic==2.8.2